requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.4
django-background-tasks==1.2.5
django-compat==1.0.15
gunicorn==19.9.0
idna==2.9
numpy==1.18.3
pandas==1.0.3
python-bittrex==0.3.0
python-dateutil==2.8.1
pytz==2019.3
requests==2.23.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
uwsgidecorators==1.1.0
